[launcher] Count buttons and create expander if greater 0
There is a fairly serious bug in moblin-panel-applications. When the
categories are assembled, it is possible for an entire category to be
skipped if certain criteria are met. It is due to a logic error.

The mnb_launcher_fill_category() function uses
launcher_button_create_from_entry() to get each icon. That function,
in turn is nice enough to return NULL if the executable is not found.

The problem is that the expander is only created if the *last* icon
retrieved is not null. So, there can be 100 valid .desktop files and
one that would normally be skipped at the end, but the entire category
is hidden!

Fixes: MB#6838